done EX

+++ Lab1_ springDataJPA/src/main/java/com/springboot/repository/CustomerRepo.java
@@ -2,9 +2,8 @@ package com.springboot.repository;
 
 import com.springboot.entity.Customer;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CustomerRepo extends JpaRepository<Customer,Integer> {
+public interface CustomerRepo extends JpaRepository<Customer, Long> {
 }
